Refactor authentication guards: rename auth.guard.ts to jwt.auth.guard.ts and update imports; remove deprecated auth.guard.ts
