+++ backend/src/auth/controllers/auth.controller.spec.ts
@@ -9,7 +9,7 @@ import { SmsModule } from 'src/sms/sms.module';
 import { UserModule } from 'src/user/user.module';
 import { BcryptService } from '../providers/bcrypt.service';
 import { CryptoService } from '../providers/crypto.service';
-import { JwtAuthGuard } from '../guards/auth.guard';
+import { JwtAuthGuard } from '../guards/jwt.auth.guard';
 import { GoogleStrategy } from '../strategies/google.strategy';
 
 describe('AuthController', () => {
@@ -19,20 +19,20 @@ describe('AuthController', () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [AuthController],
       providers: [AuthService, BcryptService, CryptoService, JwtAuthGuard, GoogleStrategy],
-  imports: [
-    UserModule,
-    SessionModule,
-    SmsModule,
-    EmailModule,
-    JwtModule.registerAsync({
-      imports: [ConfigModule],
-      inject: [ConfigService],
-      useFactory: async (configService: ConfigService) => ({
-        secret: configService.get<string>('JWT_SECRET'),
-        global: true,
-      }),
-    }),
-  ],
+      imports: [
+        UserModule,
+        SessionModule,
+        SmsModule,
+        EmailModule,
+        JwtModule.registerAsync({
+          imports: [ConfigModule],
+          inject: [ConfigService],
+          useFactory: async (configService: ConfigService) => ({
+            secret: configService.get<string>('JWT_SECRET'),
+            global: true,
+          }),
+        }),
+      ],
     }).compile();
 
     controller = module.get<AuthController>(AuthController);
